naming CoreQueryBuilder <-> CoreRequestBuilder
