+++ toto.json
@@ -1,66 +1,67 @@
-{
-  "id": "1HRXNFQyXxZwGYk",
-  "name": "group:group3",
-  "displayName": "group3",
-  "source": 2,
-  "config": 1540,
-  "description": "",
-  "settings": [],
-  "creation": 1619097660,
-  "initiator": {
-    "id": "xKrTibvwO2LAabf",
-    "circleId": "1HRXNFQyXxZwGYk",
-    "singleId": "rUMee68NHAca8UE",
-    "userId": "circles",
-    "userType": 10000,
-    "basedOn": {
-      "id": "rUMee68NHAca8UE",
-      "name": "app:circles:rUMee68NHAca8UE",
-      "displayName": "circles",
-      "source": 10001,
-      "config": 8193,
-      "description": "",
-      "settings": [],
-      "creation": 1619087109,
-      "initiator": null
-    },
-    "instance": "orange.local",
-    "local": true,
-    "level": 9,
-    "status": "Member",
-    "displayName": "circles",
-    "displayUpdate": 1619097660,
-    "note": "",
-    "contactId": "",
-    "contactMeta": "",
-    "joined": 1619097660
-  },
-  "owner": {
-    "id": "xKrTibvwO2LAabf",
-    "circleId": "1HRXNFQyXxZwGYk",
-    "singleId": "rUMee68NHAca8UE",
-    "userId": "circles",
-    "userType": 10000,
-    "basedOn": {
-      "id": "rUMee68NHAca8UE",
-      "name": "app:circles:rUMee68NHAca8UE",
-      "displayName": "circles",
-      "source": 10001,
-      "config": 8193,
-      "description": "",
-      "settings": [],
-      "creation": 1619087109,
-      "initiator": null
-    },
-    "instance": "orange.local",
-    "local": true,
-    "level": 9,
-    "status": "Member",
-    "displayName": "circles",
-    "displayUpdate": 1619097660,
-    "note": "",
-    "contactId": "",
-    "contactMeta": "",
-    "joined": 1619097660
+[
+  {
+    "mount": {
+      "id": 1,
+      "circleId": "buXblrMDt97jXWY",
+      "mountId": "",
+      "parent": 1,
+      "owner": {
+        "id": "41ba372593f8403",
+        "circleId": "buXblrMDt97jXWY",
+        "singleId": "E9wpWiPTUUMMsyU",
+        "userId": "test3",
+        "userType": 1,
+        "basedOn": null,
+        "instance": "green.local",
+        "local": false,
+        "level": 1,
+        "status": "Member",
+        "displayName": "test3",
+        "displayUpdate": 1619556617,
+        "note": "",
+        "contactId": "",
+        "contactMeta": "",
+        "joined": 1619556617
+      },
+      "initiator": {
+        "id": "bkUNjfQwIBeK5Z3",
+        "circleId": "buXblrMDt97jXWY",
+        "singleId": "WkjFRgYKdu6OYTy",
+        "userId": "cult",
+        "userType": 1,
+        "basedOn": null,
+        "instance": "orange.local",
+        "local": true,
+        "level": 9,
+        "status": "Member",
+        "displayName": "cult",
+        "displayUpdate": 1619531282,
+        "note": "",
+        "contactId": "",
+        "contactMeta": "",
+        "joined": 1619531282,
+        "inheritedBy": {
+          "id": "WkjFRgYKdu6OYTy",
+          "user_id": "cult",
+          "user_type": 1,
+          "instance": "orange.local",
+          "link": {
+            "singleId": "WkjFRgYKdu6OYTy",
+            "circleId": "buXblrMDt97jXWY",
+            "level": 9,
+            "inheritanceFirst": "WkjFRgYKdu6OYTy",
+            "inheritanceLast": "buXblrMDt97jXWY",
+            "inheritancePath": [
+              "WkjFRgYKdu6OYTy"
+            ],
+            "inheritanceDepth": 1
+          }
+        }
+      },
+      "token": "tokenrMDt97jXWY",
+      "password": "",
+      "mountPoint": "\\/test_mount\\/",
+      "mountPointHash": "sfhasdlkjfad"
+    }
   }
-}
+]
